+++ cmd/edit.go
@@ -74,6 +74,7 @@ var editCmd = &cobra.Command{
 			viper.GetBool(ALLOW_PROJECT_NAME),
 			printTimeEntryImpl(c, format, asJSON),
 			!viper.GetBool(ALLOW_INCOMPLETE),
+			true,
 		)
 	}),
 }