+++ cmd/in.go
@@ -61,6 +61,7 @@ var inCmd = &cobra.Command{
 			viper.GetBool(ALLOW_PROJECT_NAME),
 			printTimeEntryImpl(c, format, asJSON),
 			!viper.GetBool(ALLOW_INCOMPLETE),
+			true,
 		)
 	}),
 }